+++ editors/neutrino-mode.el
@@ -164,6 +164,7 @@
     "num"
     "numel"
     "ones"
+    "packeur"
     "pause"
     "phi"
     "pi"